refactor: split pages.py into PageTypeRegistry + per-type modules
Phase 1 of the page-type catalog refactor. The 774-line pages.py is replaced
by a pages/ package where each built-in page type (list, form, dashboard)
lives in its own module and self-registers via a PageTypeRegistry. The
dispatcher in __init__.py now routes by registry lookup instead of an
if/elif chain, which is what lets future phases add new page types (detail,
edit, grid, kanban, calendar, tree, settings) without touching the
dispatcher, and lets plugins register custom types via entry points.

Public API (parse_pages, find_project_root, generate_page_components) is
preserved, so cli/generate.py and the test suite work unchanged. Generator
output is byte-identical to the pre-refactor version.

Built-in emitter modules are named *_type.py rather than list.py / form.py /
dashboard.py because Python binds submodules as package attributes, which
would shadow the built-in list type inside parse_pages and break
isinstance(x, list).

Includes 10 new tests covering the registry contract; full suite is 99/99.
